+++ code/cell.modules/pkg.sys/ui.dev/src/hooks/actions/index.ts
@@ -1,1 +1,1 @@
-export * from './useActionController';
+export * from './useActionPanelController';
